feat: removing emails by record id works
